Moved crossmint minting into separate file
