+++ apps/api/src/routes/claims/index.ts
@@ -2,9 +2,9 @@ import { zValidator } from '@hono/zod-validator';
 import { Hono } from 'hono';
 import { decode, verify } from 'hono/jwt';
 import { env } from 'hono/adapter';
-import { getChainId } from '@sni/address-utils';
 import { collections } from './config';
-import { ClaimBody, JWTToken, CrossmintResponse } from './schemas';
+import { ClaimBody, JWTToken } from './schemas';
+import { mintOptimismToken } from './providers/crossmint';
 
 const app = new Hono();
 
@@ -13,8 +13,6 @@ app.post(
   zValidator('json', ClaimBody),
 
   async (c) => {
-    const { CROSSMINT_API_URL } = env<{ CROSSMINT_API_URL: string }>(c);
-    const { CROSSMINT_API_KEY } = env<{ CROSSMINT_API_KEY: string }>(c);
     const { CLAIMS_SECRET } = env<{ CLAIMS_SECRET: string }>(c);
 
     const body = ClaimBody.parse(await c.req.json());
@@ -29,54 +27,15 @@ app.post(
 
     const { payload } = JWTToken.parse(decode(token));
 
-    const collectionId: string = payload.collection;
-    const collectionConfig = collections[collectionId];
+    const collectionConfig = collections[payload.collection];
+    const network = collectionConfig.network;
 
-    const image = `${collectionConfig.metadata.imagePrefix}${payload.seed}.jpg`;
-    const metadata = {
-      name: collectionConfig.metadata.name,
-      description: collectionConfig.metadata.description,
-      image,
-    };
-
-    const mintingConfig = {
-      metadata,
-      recipient: `${collectionConfig.network}:${address}`,
-    };
-
-    const resp = await fetch(
-      `${CROSSMINT_API_URL}/${collectionId}/nfts/${payload.id}`,
-      {
-        method: 'PUT',
-        body: JSON.stringify(mintingConfig),
-        headers: {
-          'Content-Type': 'application/json',
-          'X-API-KEY': CROSSMINT_API_KEY,
-        },
-      }
-    );
-
-    const data = CrossmintResponse.parse(await resp.json());
-
-    if (data.onChain.owner && data.onChain.owner !== address) {
-      return c.json(
-        {
-          error: true,
-          message: 'Token was already claimed for different owner address',
-        },
-        400
-      );
+    switch (network) {
+      case 'optimism':
+        return mintOptimismToken(address, payload, collectionConfig, c);
+      case 'opal':
+        break;
     }
-
-    // If the token was successfully minted, return all the data plus DID address
-    if (data.onChain.status === 'success') {
-      const chainId = getChainId(data.onChain.chain);
-      const assetDID = `did:asset:eip155:${chainId}.erc721:${data.onChain.contractAddress}:${data.onChain.tokenId}`;
-
-      return c.json({ ...data, assetDID });
-    }
-
-    return c.json(data);
   }
 );
