+++ apps/api/src/routes/claims/types.ts
@@ -1,0 +1,6 @@
+//TODO: Payload type should be from ZOD schema
+export type Payload = {
+  collection: string;
+  id: string;
+  seed: number;
+};
